sessionctx: separate the slow log functionality from the session file
